+++ app/models/sla_cache_spent_query.rb
@@ -172,7 +172,7 @@ class SlaCacheSpentQuery < Query
     return @all_sla_type_values if @all_sla_type_values
 
     values ||= []
-    SlaType.pluck(:name,:id).map { |name,id|
+    SlaType.sorted.pluck(:name,:id).map { |name,id|
       values << [name.to_s,id.to_s]
     }
     @all_sla_type_values = values